// nightclub/lib/venues.generated.ts
export const regions = ["경기","대구","부산","서울","울산","인천","충남"] as const;

export const citiesByRegion: Record<string, string[]> = {
  "경기": [
    "고양",
    "파주",
    "성남",
    "수원"
  ],
  "대구": [
    "대구"
  ],
  "부산": [
    "부산"
  ],
  "서울": [
    "서울"
  ],
  "울산": [
    "울산"
  ],
  "인천": [
    "인천"
  ],
  "충남": [
    "천안"
  ]
};

export const allThemes = ["7080","DJ파티","감성주점","강남","라운지","소셜댄스","클럽"] as const;